added check IsSyncedToS3 for safety

+++ api/net/Areas/Editor/Controllers/StorageController.cs
@@ -593,8 +593,14 @@ public class StorageController : ControllerBase
                 }
                 else
                 {
-                    _fileReferenceService.DeleteAndSave(fileReference);
-                    failedUploads.Add($"{fileReference.Path} (file not found, broken reference deleted)");
+                    if (!fileReference.IsSyncedToS3)
+                    {
+                        // GetFiles method will return fileReference that is not synced to S3 normally, for extra safety, we double check here fileReference.IsSyncedToS3
+                        // if file not found, and file was not synced to S3, then it is a broken reference, we delete it
+                        _logger.LogInformation("File not found, broken reference deleted: {Path}", fileReference.Path);
+                        _fileReferenceService.DeleteAndSave(fileReference);
+                        failedUploads.Add($"{fileReference.Path} (file not found, broken reference deleted)");
+                    }
                 }
             }
             catch (Exception ex)
